手数料など反映するように編集
