+++ app/javascript/application.js
@@ -2,4 +2,5 @@
 import "@hotwired/turbo-rails"
 import "controllers"
 import "card"
+import "price_calculation"
 //= require items
