+++ config/importmap.rb
@@ -1,8 +1,9 @@
 # Pin npm packages by running ./bin/importmap
 
-pin "application"
 pin "@hotwired/turbo-rails", to: "turbo.min.js"
 pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
 pin "card", to: "card.js"
+pin "application", preload: true
+pin "price_calculation", to: "price_calculation.js"
